Move current user data source into domain package
