+++ internal/provider/currentuser/data_source_test.go
@@ -1,4 +1,4 @@
-package provider
+package currentuser
 
 import (
 	"testing"
